refactor: API 기능별 모듈 분리
- 기존 utils/api.ts 파일을 기능별로 auth.ts, question.ts, core.ts로 분리
  - 인증, 질문/답변, axios 인스턴스/인터셉터 별로 역할 분리
- 모든 API 함수의 console.error 메시지 문체 통일 (요청 실패)
- window.location.href = "/" 경로를 ROUTES.ROOT 상수로 대체
